C353652 implemented
